Update octopus-printer.cfg
I ran into an issue where I ended up with a `Move out of range: 357.500 450.000 40.000 [198.000]` error after the G32 finishes. 

This is using SuperSlicer Version 2.3.57 with the stock Voron 2.4 300mm profiles. Not exactly sure how I managed to get this to happen as I have printed with these profiles successfully just minutes before.

Seems the gcode/printer somehow ends up in relative positioning, and @chun in #slicers_and_print_help on the Voron discord suggested to add a G90 to the PRINT_START macro.
Figured I'd mention it here in case that is somehow we would want as a default in case others also run into this :-)