reduce ACC to a sane value of 3000
Reduce the default acceleration from 10000 to 3000. The name of the parameter is misleading: This used to be a maximum value, but nowadays it is just a default value. 10000 ACC as a default on a non-tuned V2.4 is a bad move. People who use Cura or other slicers without acceleration control get bad prints with this. As I wrote, this is not an upper limit. You can put 3000 in here, and if you enter 6000 in your slicer profile as printing acceleration, the 6000 are executed by the printer.